HH-81172 rename nab-starter-tests to nab-tests
