+++ nab-tests/src/test/java/ru/hh/nab/datasource/DataSourceTestConfig.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import ru.hh.metrics.StatsDSender;
-import ru.hh.nab.starter.NabTestConfig;
-import ru.hh.nab.datasource.postgres.embedded.EmbeddedPostgresDataSourceFactory;
+import ru.hh.nab.testbase.NabTestConfig;
+import ru.hh.nab.testbase.postgres.embedded.EmbeddedPostgresDataSourceFactory;
 
 import javax.sql.DataSource;
