+++ nab-tests/src/test/java/ru/hh/nab/hibernate/model/test/PackageScanEntity.java
