+++ nab-tests/src/test/java/ru/hh/nab/starter/NabApplicationTest.java
@@ -3,6 +3,7 @@ package ru.hh.nab.starter;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
+import ru.hh.nab.testbase.NabTestConfig;
 
 public class NabApplicationTest {
